Code now considers startposition when checking returned records
Previously, the code would check if maxrecords was greater than the
matched records, but did not consider that the startposition might be
non-zero.  This is now addressed for different cases of startposition,
maxrecords and matched.

Code change in csw2.py only, to be tested. The code in csw3.py also
needs modified in a similar way.