Functional test post_GetRecord-end expected XML corrected
The numberOfRecordsReturned in the expected XML for the test
default/post/GetRecords-end.xml was inconsistent with the number of
records in the expected XML.  This test failed when the code generating
the numberOfRecordsReturned attribute was updated.

Both the code and the expected XML for this test now agree and reflect
the correct value.