CSW3 functionaltests dictate that nextrecord='0' for maxrecords='0'
Although the CSW2 behaviour is for nextrecord='1', because of the policy
of the ets-csw202 test suite, the CSW3 functionaltests in the pycsw repo
prescribe that nextrecord='0'.  This has been so for over a year, so I'm
not going to change the test.  The CSW2 behaviour might still be brought
into line with CSW2, as there is no reason for it not to (apart from the
assertion of the ets-csw202 test maintainer that the standard disallows
it).